Organize CSA implementation utilities
