+++ megatron/core/transformer/experimental_attention_variant/csa.py
@@ -15,12 +15,14 @@ from megatron.core.packed_seq_params import PackedSeqParams
 from megatron.core.process_groups_config import ProcessGroupCollection
 from megatron.core.tensor_parallel.mappings import gather_from_sequence_parallel_region
 from megatron.core.transformer.enums import AttnMaskType
-from megatron.core.transformer.experimental_attention_variant import csa_cp_layout_kernels
-from megatron.core.transformer.experimental_attention_variant import csa_cp_utils as cp_utils
-from megatron.core.transformer.experimental_attention_variant.csa_fused_compressor import (
+from megatron.core.transformer.experimental_attention_variant.csa_utils import (
+    cp_layout_kernels as csa_cp_layout_kernels,
+)
+from megatron.core.transformer.experimental_attention_variant.csa_utils import cp_utils
+from megatron.core.transformer.experimental_attention_variant.csa_utils.fused_compressor import (
     maybe_compress_thd_fused,
 )
-from megatron.core.transformer.experimental_attention_variant.csa_kernels import (
+from megatron.core.transformer.experimental_attention_variant.csa_utils.kernels import (
     FusedCSAIndexerSparseAttnFromTopkFunc,
     batch_of_row,
     build_flat_topk_idxs,
@@ -203,7 +205,7 @@ def _compressed_thd_topk_to_local(
 # THD (packed) variants of the index helpers above.
 #
 # Both produce per-row local-to-segment indices in the SAME index space that
-# ``csa_kernels.local_to_global_flat(..., cu_seqlens_q=..., cu_seqlens_kv=...)``
+# ``csa_utils.kernels.local_to_global_flat(..., cu_seqlens_q=..., cu_seqlens_kv=...)``
 # expects: each row is one query token in the packed layout, each value is
 # either ``-1`` (invalid / future position) or a non-negative local KV id in
 # ``[0, seqlen_kv_full[batch_of_row])`` where ``seqlen_kv_full[b] =
@@ -1259,7 +1261,7 @@ class Compressor(MegatronModule):
         # Additive fused fast path (CuTe DSL, one kernel per direction) for the
         # gather / overlap-window / softmax / weighted-sum region below. Returns None
         # (-> keep the eager region) for any unsupported configuration; see
-        # csa_fused_compressor.py for the gating rules.
+        # csa_utils/fused_compressor.py for the gating rules.
         compressed_thd = None
         if not pre_grouped:
             compressed_thd = maybe_compress_thd_fused(
