+++ megatron/core/transformer/experimental_attention_variant/csa_utils/__init__.py
@@ -1,0 +1,3 @@
+# Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+
+"""Implementation utilities for compressed sparse attention."""
